fix bug

+++ sampling.py
@@ -9,8 +9,8 @@ def sampling_process(
         std_arrival_time: float,
         num_process: int
     ) -> list[tuple]:
-    burst_time_array = np.random.normal(loc=mean_burst_time,scale=std_burst_time,size=num_process)
-    arrival_time_array = np.random.normal(loc=mean_arrival_time,scale=std_arrival_time,size=num_process)
+    burst_time_array = np.int32(np.random.normal(loc=mean_burst_time,scale=std_burst_time,size=num_process))
+    arrival_time_array = np.int32(np.random.normal(loc=mean_arrival_time,scale=std_arrival_time,size=num_process))
     result_sampling = []
     for i in range(num_process):
         result_sampling.append((i,arrival_time_array[i],burst_time_array[i]))
